+++ helm/private/BUILD.bazel
@@ -5,13 +5,11 @@ stamp_build_setting(name = "stamp")
 filegroup(
     name = "bzl_srcs",
     srcs = glob(["**/*.bzl"]) + [
-        "//helm/private/installer:bzl_srcs",
         "//helm/private/packager:bzl_srcs",
         "//helm/private/pusher:bzl_srcs",
         "//helm/private/registrar:bzl_srcs",
-        "//helm/private/stamper:bzl_srcs",
-        "//helm/private/uninstaller:bzl_srcs",
-        "//helm/private/upgrader:bzl_srcs",
+        "//helm/private/runner:bzl_srcs",
+        "//helm/private/stamp:bzl_srcs",
     ],
     visibility = ["//:__subpackages__"],
 )
