--- helm/private/installer/helm_install.bat.template
@@ -1,11 +1,0 @@
-ECHO OFF
-
-{image_pushers}
-
-set "helm_args="
-for /f "delims=" %%a in ({helm_opts}) do set helm_args=!helm_args! %%a
-
-set "install_args="
-for /f "delims=" %%a in ({install_opts}) do set install_args=!install_args! %%a
-
-{helm} !helm_args! install !install_args! {install_name} {chart} %@%
