--- helm/private/installer/helm_install.sh.template
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-
-set -euxo pipefail
-
-{image_pushers}
-
-eval exec {helm} $(cat {helm_opts}) install $(cat {install_opts}) {install_name} {chart} "$@"
