+++ helm/private/packager/BUILD.bazel
@@ -4,7 +4,9 @@ go_binary(
     name = "packager",
     srcs = ["packager.go"],
     visibility = ["//visibility:public"],
-    deps = ["@go_yaml_yaml"],
+    deps = [
+        "@go_yaml_yaml",
+    ],
 )
 
 filegroup(
