--- helm/private/reinstaller/helm_reinstall.sh.template
@@ -1,8 +1,0 @@
-#!/usr/bin/env bash
-
-set -euo pipefail
-
-{image_pushers}
-
-eval {helm} $(cat {helm_opts}) uninstall $(cat {reinstall_opts}) {install_name}
-eval exec {helm} $(cat {helm_opts}) install $(cat {reinstall_opts}) {install_name} {chart} "$@"
