+++ helm/private/runner/runner_wrapper.bat.template
@@ -1,0 +1,9 @@
+ECHO OFF
+
+set HELM_BIN="{HELM_BIN}"
+set STABLE_STATUS_FILE="{STABLE_STATUS_FILE}"
+set VOLATILE_STATUS_FILE="{VOLATILE_STATUS_FILE}"
+
+{EXTRA_CMDS}
+
+{RUNNER} {HELM_OPTS} %@%
