+++ helm/private/stamp/BUILD.bazel
@@ -1,8 +1,9 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
-go_binary(
-    name = "stamper",
-    srcs = ["stamper.go"],
+go_library(
+    name = "stamp",
+    srcs = ["stamp.go"],
+    importpath = "github.com/abrisco/rules_helm/helm/private/stamp",
     visibility = ["//visibility:public"],
     deps = ["@go_yaml_yaml"],
 )
