--- helm/private/uninstaller/BUILD.bazel
@@ -1,28 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-exports_files([
-    "helm_uninstall.bat.template",
-    "helm_uninstall.sh.template",
-])
-
-config_setting(
-    name = "windows",
-    constraint_values = ["@platforms//os:windows"],
-)
-
-alias(
-    name = "template",
-    actual = select({
-        ":windows": ":helm_uninstall.bat.template",
-        "//conditions:default": ":helm_uninstall.sh.template",
-    }),
-)
-
-filegroup(
-    name = "bzl_srcs",
-    srcs = glob(
-        ["**/*.bzl"],
-        allow_empty = True,
-    ),
-    visibility = ["//:__subpackages__"],
-)
