--- helm/private/upgrader/helm_upgrade.bat.template
@@ -1,11 +1,0 @@
-ECHO OFF
-
-{image_pushers}
-
-set "helm_args="
-for /f "delims=" %%a in ({helm_opts}) do set helm_args=!helm_args! %%a
-
-set "upgrade_args="
-for /f "delims=" %%a in ({upgrade_opts}) do set upgrade_args=!upgrade_args! %%a
-
-{helm} !helm_args! upgrade !upgrade_args! {install_name} {chart} %@%
