--- helm/private/upgrader/helm_upgrade.sh.template
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-
-set -euxo pipefail
-
-{image_pushers}
-
-eval exec {helm} $(cat {helm_opts}) upgrade $(cat {upgrade_opts}) {install_name} {chart} "$@"
